Move Node classmethods to node_type_registry.
