+++ pydiggy/__init__.py
@@ -6,7 +6,8 @@ __author__ = """Adam Hopkins"""
 __email__ = "admhpkns@gmail.com"
 __version__ = "0.1.0"
 
-from pydiggy.node import Facets, Node, get_node, is_facets
+from pydiggy.node import Facets, Node, is_facets
+from pydiggy.node_type_registry import get_node_type, NodeTypeRegistry
 from pydiggy.operations import generate_mutation, hydrate, query, run_mutation
 from pydiggy._types import count, exact, geo, index, lang, reverse, uid, upsert
 
@@ -17,12 +18,13 @@ __all__ = (
     "Facets",
     "generate_mutation",
     "geo",
-    "get_node",
+    "get_node_type",
     "hydrate",
     "is_facets",
     "index",
     "lang",
     "Node",
+    "NodeTypeRegistry",
     "query",
     "reverse",
     "run_mutation",
