+++ pydiggy/node.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 
 import copy
-import inspect
 import json
+import inspect
 from collections import namedtuple
 from copy import deepcopy
 from dataclasses import dataclass, field
@@ -89,19 +89,6 @@ def _force_instance(
     return directive(*args)
 
 
-def get_node(name: str) -> Node:
-    """
-    Retrieve a registered node class.
-
-    Example: Region = get_node("Region")
-
-    This is a safe method to make sure that any models used have been
-    declared as a Node.
-    """
-    registered = {x.__name__: x for x in Node._nodes}
-    return registered.get(name, None)
-
-
 class NodeMeta(type):
     def __new__(cls, name, bases, attrs, **kwargs):
         directives = [
@@ -136,24 +123,42 @@ class NodeMeta(type):
 
 
 class Node(metaclass=NodeMeta):
-    uid: int
+    _instances = dict()
+    _staged = dict()
 
-    _i = _count()
-    _nodes = []
-    _staged = {}
+
+    @classmethod
+    def _get_staged(cls):
+        return cls._staged
+
+    @classmethod
+    def _clear_staged(cls):
+        cls._staged = {}
+        # NodeTypeRegistry._i needs to be updated when clearing staged nodes
+
+    @classmethod
+    def _reset(cls) -> None:
+        cls._instances = dict()
+
+    @classmethod
+    def _get_name(cls) -> str:
+        return cls.__name__
 
     def __init_subclass__(cls, is_abstract: bool = False) -> None:
         if not is_abstract:
-            cls._register_node(cls)
+            from pydiggy.node_type_registry import NodeTypeRegistry
+            NodeTypeRegistry._register_node_type(cls)
 
     def __init__(self, uid=None, **kwargs):
+
+        from pydiggy.node_type_registry import NodeTypeRegistry
         if uid is None:
             # TODO:
             # - There probably should be another property that is set here
             #   so that it is possible to identify with a boolean if the instance
             #   is brand new (and has never been committed to the DB) or if it
             #   is being freshly generated
-            uid = next(self._generate_uid())
+            uid = next(NodeTypeRegistry._generate_uid())
 
         self.uid = uid
         self._dirty = set()
@@ -162,7 +167,7 @@ class Node(metaclass=NodeMeta):
         # - perhaps this code to generate self._annotations belongs somewhere
         #   else. Regardless, there is a lot of cleanup in this module that
         #   could probably use this property instead of running get_type_hints
-        localns = {x.__name__: x for x in Node._nodes}
+        localns = {x.__name__: x for x in NodeTypeRegistry._node_types}
         localns.update({"List": List, "Union": Union, "Tuple": Tuple})
         self._annotations = get_type_hints(
             self.__class__, globalns=globals(), localns=localns
@@ -196,7 +201,8 @@ class Node(metaclass=NodeMeta):
         return hash(self.uid)
 
     def __getattr__(self, attribute):
-        if attribute in self.__annotations__:
+        if (not attribute == "__annotations__")\
+            and (attribute in self.__annotations__):
             raise MissingAttribute(self, attribute)
         super().__getattribute__(attribute)
 
@@ -211,7 +217,10 @@ class Node(metaclass=NodeMeta):
 
         orig = self.__dict__.get(name, None)
         self.__dict__[name] = value
-        if hasattr(self, "_init") and self._init and not name.startswith("_"):
+        # TODO: This causes issues with Node types that want private variables
+        if hasattr(self, "_init")\
+            and self._init\
+            and not name.startswith("_"):
             self._dirty.add(name)
         if name in self._directives and any(
             isinstance(d, reverse) for d in self._directives[name]
@@ -258,301 +267,6 @@ class Node(metaclass=NodeMeta):
                         orig, reverse_name, self, directive.many, remove=True
                     )
 
-    @classmethod
-    def _reset(cls) -> None:
-        cls._i = _count()
-        cls._instances = dict()
-
-    @classmethod
-    def _register_node(cls, node: Node) -> None:
-        cls._nodes.append(node)
-
-    @classmethod
-    def _get_name(cls) -> str:
-        return cls.__name__
-
-    @classmethod
-    def _generate_schema(cls) -> str:
-        nodes = cls._nodes
-        edges = {}
-        schema = []
-        type_schema = []
-        edge_schema = []
-        unknown_schema = []
-
-        type_schema.append(f"_type: string .")
-
-        # TODO:
-        # - Prime candidate for some refactoring to reduce complexity
-        for node in nodes:
-            name = node._get_name()
-            type_schema.append(f"{name}: bool @index(bool) .")
-
-            annotations = get_type_hints(node)
-            for prop_name, prop_type in annotations.items():
-                # TODO:
-                # - Probably need to be a little more careful for the origins
-                # - Currently, it is assuming Union[something, None],
-                #   but if you had two legit items inside Union (or anything else)
-                #   then the second would be ignored. Which, is probably correct
-                #   unless Dgraph schema would not allow multiple types for a single
-                #   predicate. If it is possible, then the solution may simply
-                #   be to loop over a list of deepcopy(annotations.items()),
-                #   and append all the __args__ to that list to extend the iteration
-                # - is_list_type should probably become its own function
-                is_list_type = (
-                    True
-                    if isinstance(prop_type, _GenericAlias)
-                    and prop_type.__origin__ in (list, tuple)
-                    else False
-                )
-
-                if (
-                    isinstance(prop_type, _GenericAlias)
-                    and prop_type.__origin__ in ACCEPTABLE_GENERIC_ALIASES
-                ):
-                    prop_type = prop_type.__args__[0]
-
-                prop_type = PropType(
-                    prop_type,
-                    is_list_type,
-                    node._directives.get(prop_name, []),
-                )
-
-                if prop_name in edges:
-                    if prop_type != edges.get(
-                        prop_name
-                    ) and not cls._is_node_type(prop_type[0]):
-
-                        # Check if there is a type conflict
-                        if (
-                            edges.get(prop_name).directives
-                            != prop_type.directives
-                            and all(
-                                (
-                                    inspect.isclass(x)
-                                    and issubclass(x, Directive)
-                                )
-                                or issubclass(x.__class__, Directive)
-                                for x in edges.get(prop_name).directives
-                            )
-                            and all(
-                                (
-                                    inspect.isclass(x)
-                                    and issubclass(x, Directive)
-                                )
-                                or issubclass(x.__class__, Directive)
-                                for x in prop_type.directives
-                            )
-                        ):
-                            pass
-                        else:
-                            raise ConflictingType(
-                                prop_name, prop_type, edges.get(prop_name)
-                            )
-
-                # Set the type for translating the value
-                if prop_type[0] in ACCEPTABLE_TRANSLATIONS:
-                    edges[prop_name] = prop_type
-                elif cls._is_node_type(prop_type[0]):
-                    edges[prop_name] = PropType(
-                        "uid",
-                        is_list_type,
-                        node._directives.get(prop_name, []),
-                    )
-                else:
-                    if prop_name != "uid":
-                        origin = getattr(prop_type[0], "__origin__", None)
-                        # if origin and origin
-                        unknown_schema.append(
-                            f"{prop_name}: {prop_type[0]} || {origin}"
-                        )
-
-        # TODO:
-        # - When the v1.1 comes out, will need to address this:
-        #
-        #         for edge_name, (edge_type, is_list_type) in edges.items():
-        #             type_name = cls._get_type_name(edge_type)
-        #             # Currently, Dgraph does not support [uid] schema.
-        #             # See https://github.com/dgraph-io/dgraph/issues/2511
-        #             if is_list_type and type_name != 'uid':
-        for edge_name, edge in edges.items():
-            type_name = cls._get_type_name(edge.prop_type)
-            if edge.is_list_type and type_name != "uid":
-                type_name = f"[{type_name}]"
-
-            directives = edge.directives
-            directives = " ".join([str(d) for d in directives] + [""])
-
-            edge_schema.append(f"{edge_name}: {type_name} {directives}.")
-
-        type_schema.sort()
-        edge_schema.sort()
-
-        schema = type_schema + edge_schema
-
-        return "\n".join(schema), "\n".join(unknown_schema)
-
-    @classmethod
-    def _get_type_name(cls, schema_type):
-        if isinstance(schema_type, str):
-            return schema_type
-
-        name = schema_type.__name__
-        if cls._is_node_type(schema_type):
-            return name
-        else:
-            if name not in DGRAPH_TYPES:
-                raise Exception(f"Could not find type: {name}")
-            return DGRAPH_TYPES.get(name)
-
-    @classmethod
-    def _get_staged(cls):
-        return cls._staged
-
-    @classmethod
-    def _clear_staged(cls):
-        cls._staged = {}
-        cls._i = _count()
-
-    @classmethod
-    def _hydrate(
-        cls, raw: Dict[str, Any], types: Dict[str, Node] = None
-    ) -> Node:
-        # TODO:
-        # - Accept types that are passed. Loop thru them and register if needed
-        #   and raising an exception if they are not valid.
-        # - This method is another candidate for some refactoring to reduce
-        #   complexity.
-        # - Should create a Facets type so that the type annotation of this function
-        #   is _hydrate(cls, raw: str, types: Dict[str, Node] = None) -> Union[Node, Facets]
-        registered = {x.__name__: x for x in Node._nodes}
-        localns = {x.__name__: x for x in Node._nodes}
-        localns.update({"List": List, "Union": Union, "Tuple": Tuple})
-
-        if "_type" in raw and raw.get("_type") in registered:
-            if "uid" not in raw:
-                raise InvalidData("Missing uid.")
-
-            k = registered[raw.get("_type")]
-
-            keys = deepcopy(list(raw.keys()))
-            facet_data = [
-                (key.split("|")[1], raw.pop(key)) for key in keys if "|" in key
-            ]
-
-            kwargs = {"uid": int(raw.pop("uid"), 16)}
-            delay = []
-            computed = {}
-
-            pred_items = [
-                (pred, value)
-                for pred, value in raw.items()
-                if not pred.startswith("_")
-            ]
-
-            annotations = get_type_hints(
-                k, globalns=globals(), localns=localns
-            )
-            for pred, value in pred_items:
-                """
-                The pred falls into one of three categories:
-                1. predicates that have already been defined
-                2. predicates that are a reverse of a relationship
-                3. predicates that are used for some computed value
-                """
-                if pred in annotations:
-                    if isinstance(value, list):
-                        prop_type = annotations[pred]
-                        is_list_type = (
-                            True
-                            if isinstance(prop_type, _GenericAlias)
-                            and prop_type.__origin__ in (list, tuple)
-                            else False
-                        )
-                        if is_list_type:
-                            value = [cls._hydrate(x) for x in value]
-                        else:
-                            if len(value) > 1:
-                                # This should NOT happen. Because uid
-                                # in dgraph is not forced 1:1 with a uid predicate
-                                # it should return as a [<Node>] with only
-                                # a single item in it. If the developer wants
-                                # multiple: then the Node definition should
-                                # be List[MyNode]
-                                # Will probably need to be revisited when
-                                # Dgraph v. 1.1 is released
-                                raise Exception("Unknown data")
-                            node = get_node(value[0].get("_type"))
-                            value = node._hydrate(value[0])
-                    elif isinstance(value, dict):
-                        value = cls._hydrate(value)
-
-                    if value is not None:
-                        kwargs.update({pred: value})
-                elif pred.startswith("~"):
-                    p = pred[1:]
-                    if isinstance(value, list):
-                        for x in value:
-                            keys = deepcopy(list(x.keys()))
-                            value_facet_data = [
-                                (k.split("|")[1], x.pop(k))
-                                for k in keys
-                                if "|" in k
-                            ]
-                            item = get_node(x.get("_type"))._hydrate(x)
-
-                            if value_facet_data:
-                                item = Facets(item, **dict(value_facet_data))
-                            delay.append((item, p, value_facet_data))
-                    elif isinstance(value, dict):
-                        delay.append(
-                            (
-                                get_node(value.get("_type"))._hydrate(value),
-                                p,
-                                None,
-                            )
-                        )
-                else:
-                    if pred.endswith("_uid"):
-                        value = int(value, 16)
-                    computed.update({pred: value})
-
-            instance = k(**kwargs)
-            for d, p, v in delay:
-                if is_facets(d):
-                    f = Facets(instance, **dict(v))
-                    setattr(d.obj, p, f)
-                else:
-                    setattr(d, p, instance)
-
-            if computed:
-                instance.computed = Computed(**computed)
-
-            if facet_data:
-                facets = Facets(instance, **dict(facet_data))
-                return facets
-            else:
-                return instance
-        return None
-
-    @classmethod
-    def json(cls) -> Dict[str, List[Node]]:
-        """
-        Return mapping of Node names to a list of node instances.
-
-        Can be used as a way to dump what node instances are in memory.
-        """
-        # TODO:
-        # - Probably should be renamed
-        # - Instrad of being a List[Node], it should probably be a Set
-        return {
-            x.__name__: list(
-                map(partial(cls._explode, max_depth=0), x._instances.values())
-            )
-            for x in cls._nodes
-            if len(x._instances) > 0
-        }
 
     @classmethod
     def _explode(
@@ -641,25 +355,6 @@ class Node(metaclass=NodeMeta):
                         item.update({"is_facets": True})
                         obj[key] = value._asdict()
         return obj
-
-    @classmethod
-    def create(cls, **kwargs) -> Node:
-        """
-        Constructor for creating a node.
-        """
-        instance = cls()
-        for k, v in kwargs.items():
-            setattr(instance, k, v)
-        return instance
-
-    @staticmethod
-    def _is_node_type(cls) -> bool:
-        """Check if a class is a <class 'Node'>"""
-        return inspect.isclass(cls) and issubclass(cls, Node)
-
-    def _generate_uid(self) -> str:
-        i = next(self._i)
-        yield f"unsaved.{i}"
 
     def to_json(self, include: List[str] = None) -> Dict[str, Any]:
         # TODO:
