+++ pydiggy/operations.py
@@ -7,6 +7,7 @@ from typing import List, Tuple, Union, get_type_hints, Dict, Any
 from pydiggy.connection import get_client, PyDiggyClient
 from pydiggy.exceptions import NotStaged
 from pydiggy.node import Node
+from pydiggy.node_type_registry import NodeTypeRegistry
 from pydiggy._types import *  # noqa
 from pydiggy.utils import _parse_subject, _raw_value
 
@@ -25,9 +26,9 @@ def _make_obj(node, pred, obj):
     # TODO:
     # - integreate utils._rdf_value
     try:
-        if Node._is_node_type(obj.__class__):
+        if NodeTypeRegistry._is_node_type(obj.__class__):
             uid, passed = _parse_subject(obj.uid)
-            staged = Node._get_staged()
+            staged = NodeTypeRegistry._get_staged()
 
             if (
                 uid not in staged
@@ -131,14 +132,14 @@ def hydrate(data: str, types: List[Node] = None) -> Dict[str, List[Node]]:
 
     output = {}
     # data = data.get(data_set)
-    registered = {x.__name__: x for x in Node._nodes}
+    registered = {x.__name__: x for x in NodeTypeRegistry._node_types}
 
     for func_name, raw_data in data.items():
         hydrated = []
         for raw in raw_data:
             if "_type" in raw and raw.get("_type") in registered:
                 cls = registered.get(raw.get("_type"))
-                hydrated.append(cls._hydrate(raw, types=types))
+                hydrated.append(NodeTypeRegistry._hydrate(raw, types=types))
 
         output[func_name] = hydrated
